Refactor: Navigation Components
